fixed tests and linter
