+++ src/registrar/templates/django/admin/domain_change_form.html
@@ -10,7 +10,7 @@
     <div class="submit-row">
         {% if original.state == original.State.READY %}
           <input type="submit" value="Place hold" name="_place_client_hold">
-        {% elif original.state == original.State.ONHOLD %}
+        {% elif original.state == original.State.ON_HOLD %}
           <input type="submit" value="Remove hold" name="_remove_client_hold">
         {% endif %}
         <input id="manageDomainSubmitButton" type="submit" value="Manage Domain" name="_edit_domain">
